separate sparse analysis into analyzer files
